Rename lib/backend.js to backend-request Remove unused middleware, remove comment
