--- src/app/lib/middleware/login-check.js
@@ -1,11 +1,0 @@
-module.exports = function( req, res, next ){
-
-	if( req.cookies.sessionid ){
-
-		next();
-
-	} else {
-
-		res.redirect( '/login/' );
-	}
-};
